requirements.txt
coverage==4.5.1
Django==2.0.2
django-bootstrap3==9.1.0
django-recaptcha==1.4.0
lml==0.0.1
mysqlclient==1.3.12
pytz==2018.3
